Move keyboard code to dedicated directory
