+++ src/device/keyboard/mod.rs
@@ -1,0 +1,5 @@
+pub mod keyboard;
+pub mod ps2_keyboard;
+
+pub use self::keyboard::*;
+pub use self::ps2_keyboard::*;
